// future-backend/src/modules/trade/dto/get-fills.dto.ts
export class FillDto {
  id: number | any;
  orderId: number | any;
  accountId: number | any;
  fee: string | any;
  quantity: string | any;
  price: string | any;
  symbol: string | any;
  liquidity: "Taker" | "Maker" | any;
  orderType: string | any;
  tradeSide: "BUY" | "SELL" | any;
  createdAt: Date | any;
}